requirements.txt
requests
numpy
Pillow
torch
transformers
torchvision
realesrgan
scipy
